Move tab components into their own folder
