+++ front/app/containers/Admin/users/tabs/Admins.tsx
@@ -5,9 +5,9 @@ import styled from 'styled-components';
 
 import SeatInfo from 'components/admin/SeatBasedBilling/SeatInfo';
 
-import UserManager from './_shared/UserManager';
-import UsersHeader from './_shared/UsersHeader';
-import messages from './messages';
+import UserManager from '../_shared/UserManager';
+import UsersHeader from '../_shared/UsersHeader';
+import messages from '../messages';
 
 const StyledBox = styled(Box)`
   display: grid;
