+++ front/app/containers/Admin/users/tabs/AllUsers.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import { Box } from '@citizenlab/cl2-component-library';
 
-import messages from './messages';
-import UserManager from './_shared/UserManager';
-import UsersHeader from './_shared/UsersHeader';
+import UserManager from '../_shared/UserManager';
+import UsersHeader from '../_shared/UsersHeader';
+import messages from '../messages';
 
 const AllUsers = () => (
   <Box className="intercom-users-registered-users">
