+++ front/app/containers/Admin/users/tabs/BannedEmails.tsx
@@ -15,8 +15,8 @@ import useUnbanEmail from 'api/email_bans/useUnbanEmail';
 
 import { useIntl } from 'utils/cl-intl';
 
-import messages from './messages';
-import UsersHeader from './_shared/UsersHeader';
+import UsersHeader from '../_shared/UsersHeader';
+import messages from '../messages';
 
 const BannedEmails = () => {
   const { formatMessage } = useIntl();
