+++ front/app/containers/Admin/users/tabs/ProjectModerators.tsx
@@ -5,9 +5,9 @@ import styled from 'styled-components';
 
 import SeatInfo from 'components/admin/SeatBasedBilling/SeatInfo';
 
-import messages from './messages';
-import UserManager from './_shared/UserManager';
-import UsersHeader from './_shared/UsersHeader';
+import UserManager from '../_shared/UserManager';
+import UsersHeader from '../_shared/UsersHeader';
+import messages from '../messages';
 
 const StyledBox = styled(Box)`
   display: grid;
@@ -15,14 +15,14 @@ const StyledBox = styled(Box)`
   grid-auto-rows: 1fr;
 `;
 
-const FolderModerators = () => (
+const ProjectModerators = () => (
   <>
-    <UsersHeader title={messages.folderManagers} />
-    <UserManager folderModeratorsOnly notCitizenlabMember includeInactive />
+    <UsersHeader title={messages.projectManagers} />
+    <UserManager projectModeratorsOnly notCitizenlabMember includeInactive />
     <StyledBox mt="20px">
       <SeatInfo seatType="moderator" />
     </StyledBox>
   </>
 );
 
-export default FolderModerators;
+export default ProjectModerators;
